Create Player Mini-App
User prompted to enter in a roster of players along with a rating for each player, and then manipulate a roster of players using the coded functions of the player roster mini app. The functions include add player, output roster, update player rating, delete player, and output players above a certain rating.